requirements.txt
sentence-transformers==2.2.2
ctransformers==0.2.5
pinecone[grpc]
langchain-pinecone==0.1.2
langchain==0.0.225
flask==3.1.0
pypdf==5.2.0
python-dotenv==1.0.1
